+++ src/main/projectiles/Needle.java
@@ -11,7 +11,7 @@ import static main.Main.*;
 
 public class Needle extends Projectile {
 
-    public Needle(PApplet p, float x, float y, float angle, Turret turret, int damage, int effectLevel, int effectDuration) {
+    public Needle(PApplet p, float x, float y, float angle, Turret turret, int damage, int effectLevel, float effectDuration) {
         super(p, x, y, angle, turret);
         position = new PVector(x, y);
         size = new PVector(2, 17);
